Extract ScreenHeader
